cornershop-backend/requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
app==0.0.1
asgiref==3.4.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
Cerberus==1.3.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==5.5
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
environs==9.3.3
Flask==2.0.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
marshmallow==3.13.0
modules==1.0.0
multidict==5.1.0
MyApplication==0.1.0
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
six==1.16.0
slackclient==2.9.3
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3